refactor(admin): the graph becomes a module with the spine as a parameter
Step 1 of the three in #1847: module → full-screen modal → only then delete
the design sections. The prototype in #1848 resolved neighbours inside
`GET /admin/designs/:id/graph`, which is the wrong home for it — every spine
wants this (Partner is the biggest node at 18 links), and the MCP
`get_entity_neighbours` tool wants the SAME resolver. Two homes for one set of
rules is the failure this codebase keeps paying for.

## Server

`src/lib/graph/` — types, a `GraphBuilder`, and a spine registry.
`resolveGraph(scope, spine, id)` is the one entry point both surfaces call.
The design spine moves to `src/lib/graph/spines/design/` with its node and
edge construction VERBATIM; `absence.ts` and its 25 tests move with it
(`git mv`, so history follows).

`GET /admin/graph/:spine/:id` replaces the design-specific route, which is
deleted rather than aliased — a second URL is a second thing to keep in step.
An unknown spine is a 404 naming the known keys, never an empty graph: empty
is indistinguishable from "no neighbours", which is the confusion this view
exists to remove.

The builder exists for one invariant: a node and its edge are added together.
The prototype pushed to two arrays by hand at nine call sites, where a node
without its edge renders as an orphan box and a summary counted off the edges
disagrees with the screen. It is now unrepresentable.

## Admin

`EntityGraph` (`components/graph/`) is the canvas, spine-agnostic — the centre
node, its label and every neighbour come from the resolver, so a new spine is
a prop change. `DesignGraphSection` is now nine lines that pass
`spine="design"`. `useEntityGraph(spine, id)` replaces `useDesignGraph`.

🔴 SVG marker ids are per-instance (`useId`). Step 2 renders this component in
a modal over a page that may already hold one, and duplicate DOM ids make the
second graph borrow the first's arrowheads.

## Verification

Behaviour-preserving, checked rather than asserted: the handler was driven
through a real container over 10 local designs (varied shapes — the two-run
`01KV3YRS…` case with `absent=2`, a bare design with one node) and the JSON
dumped before and after. Byte-identical, 41,684 bytes both runs. 25 absence
tests green in their new location; unknown-spine path exercised; no new tsc
errors in any touched file, with the new files confirmed present in the
program via --listFiles.

Refs #1847 #1848
