+++ apps/backend/src/admin/hooks/api/designs.ts
@@ -1573,59 +1573,3 @@ export const useCommitConsumption = (
  * edge" that a list of rows can never show.
  * ------------------------------------------------------------------------ */
 
-export type DesignGraphEdgeState = "present" | "derived" | "absent"
-
-export type DesignGraphProp = { key: string; value: string }
-
-export type DesignGraphNode = {
-  key: string
-  type: string
-  label: string
-  sublabel: string | null
-  state: DesignGraphEdgeState
-  count: number
-  status: string | null
-  href: string | null
-  props: DesignGraphProp[]
-  action: { label: string; href: string | null } | null
-}
-
-export type DesignGraphEdge = {
-  from: string
-  to: string
-  label: string
-  state: DesignGraphEdgeState
-  reason: string | null
-}
-
-export type AdminDesignGraphResponse = {
-  graph: {
-    spine: DesignGraphNode
-    nodes: DesignGraphNode[]
-    edges: DesignGraphEdge[]
-    summary: { links: number; absent: number; derived: number }
-  }
-}
-
-export const useDesignGraph = (
-  id: string,
-  options?: Omit<
-    UseQueryOptions<
-      AdminDesignGraphResponse,
-      FetchError,
-      AdminDesignGraphResponse,
-      QueryKey
-    >,
-    "queryFn" | "queryKey"
-  >,
-) => {
-  const { data, ...rest } = useQuery({
-    queryKey: [...designQueryKeys.detail(id), "graph"],
-    queryFn: async () =>
-      sdk.client.fetch<AdminDesignGraphResponse>(`/admin/designs/${id}/graph`, {
-        method: "GET",
-      }),
-    ...options,
-  })
-  return { ...data, ...rest }
-}
