+++ apps/backend/src/lib/graph/spines/design/absence.ts
